pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "waymo-sim"
version = "0.1.0"
description = "2D autonomous driving simulation environment with gymnasium interface"
requires-python = ">=3.9"
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "pygame>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["engine*", "agent*", "scenarios*", "viewer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
